requirements.txt
# Core
anthropic>=0.27.0
python-dotenv>=1.0.0
httpx>=0.27.0

# Quantum simulation
qiskit>=1.0.0
qiskit-aer>=0.14.0

# JSON repair fallback (P1-2)
json-repair>=0.25.0

# Evaluation / reporting (P5-2)
scipy>=1.12.0

# Reproducibility
numpy>=1.26.0

# Testing
pytest>=8.0.0
pytest-asyncio>=0.23.0
